refactor: Extract ImageController from ImageAnnotator (Phase 3b)
Second half of Phase 3 in the modular refactoring. Completes the work
started by ProjectController (PR #7) by giving image and slice
loading / navigation its own home.

New controllers/image_controller.py (869 lines) — ImageController(QObject)
plus the DimensionDialog class (moved out of annotator_window.py top
level since it is only used by process_multidimensional_image).

Methods relocated to ImageController:

Loaders:
- load_image (extension dispatch)
- load_tiff (with axes-hint detection from TIFF series metadata —
  the bugfix from arc42 for ndim>=5 TZCYX inputs is preserved
  verbatim)
- load_czi
- load_regular_image
- load_multi_slice_image (orchestrates TIFF/CZI multi-slice loading)
- process_multidimensional_image (dimension assignment dialog flow)
- create_slices (per-axis slicing + progress dialog)

Navigation / activation:
- switch_image, switch_slice, activate_slice, activate_current_slice
- update_image_list, update_slice_list, clear_slice_list
- add_slice_to_list, setup_slice_list

Image lifecycle:
- open_images, add_images_to_list, update_all_images
- remove_image, delete_selected_image, redefine_dimensions
- display_image
- is_multi_dimensional

State (current_image, current_slice, slices, image_paths, image_slices,
image_dimensions, image_shapes, all_images, image_file_name, etc.)
deliberately STAYS on the main window — accessed via self.mw.X. Same
deferral as Phase 3a; future phase may migrate ownership.

ImageAnnotator now holds self.image_controller = ImageController(self)
and keeps thin delegating methods for each so menu actions and signal
connections continue to work unchanged.

DimensionDialog moved with the methods that use it. The pre-fill from
TIFF series axes and the ndim>=7 fallback ("T"*N + ["H","W"]) are
preserved verbatim — those are the fixes from arc42's "ndim≥5
[-ndim:] slice bug" note.

Dead imports of czifile.CziFile and tifffile.TiffFile removed from
annotator_window.py since their only consumers moved into the
controller.

annotator_window.py: 4,944 → 4,122 lines (-822, -16.6%).
Total reduction vs original master: 5,900 → 4,122 (-30.1%).

Tests: all 94 pass (47 unit + 18 integration + 29 smoke).
